+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Nothing yet :)
+### Fixed
+- Adding in a guard so that refreshRate for accessing the API cannot be set lower than 30 seconds in the config.
 
 ## 1.2.0 - 2019-11-15
 ### Changed